Im AuthenticationResponse wird die UserId jetzt mitgeschickt. UserId Datentyp von Integer auf Long geändert.

+++ src/main/java/com/kayikci/learningplatform/repository/ExamRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 
 public interface ExamRepository extends JpaRepository<Exam, Long> {
-    Iterable<Exam> findByUserId(Integer userId);
+    Iterable<Exam> findByUserId(Long userId);
 
-    Optional<Exam> findByIdAndUserId(Long id, Integer userId);
+    Optional<Exam> findByIdAndUserId(Long id, Long userId);
 }
